@@ -25,6 +25,9 @@
     "@angular/router": "^14.2.4",
     "@auth0/auth0-angular": "^1.11.0",
     "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+    "@ngrx/effects": "^14.3.3",
+    "@ngrx/store": "^14.3.3",
+    "@ngrx/store-devtools": "^14.3.3",
     "@rxweb/reactive-form-validators": "^2.1.7",
     "bootstrap": "^4.5.2",
     "chart.js": "^2.9.4",
@@ -41,6 +44,7 @@
     "@angular-devkit/build-angular": "^14.2.4",
     "@angular/cli": "^14.2.4",
     "@angular/compiler-cli": "^14.2.4",
+    "@ngrx/schematics": "^15.2.1",
     "@schematics/angular": "^10.0.8",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
@@ -56,6 +60,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslib": "^2.4.0",
-    "typescript": "~4.6.4"
+    "typescript": "~4.8.2"
   }
 }
