@@ -21,6 +21,8 @@ import { MatTooltipModule } from "@angular/material/tooltip";
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AuthModule, AuthService } from "@auth0/auth0-angular";
+import { StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { environment } from "src/environments/environment";
 import { AppInitService } from "./app-init.service";
 import { AppRoutingModule } from './app-routing.module';
@@ -53,7 +55,10 @@ import { UserComponent } from './components/user/user.component';
 import { UserModule } from "./components/user/user.module";
 import { HttpAuthInterceptor } from "./helpers/interceptors/http-auth.interceptor";
 import { HttpErrorInterceptor } from "./helpers/interceptors/http-error.interceptor";
+import { metaReducers, reducers } from './store';
 import { getPolishPaginatorIntl } from "./translations/polish-paginator-intl";
+import { EffectsModule } from '@ngrx/effects';
+import { LeagueEffects } from './store/effects/league.effects';
 
 registerLocaleData(localePl);
 
@@ -115,7 +120,21 @@ registerLocaleData(localePl);
       httpInterceptor: {
         allowedList: [`${environment.api.url}/api/messages/admin`],
       }
-    })
+    }),
+    StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production }),
+    StoreModule.forRoot(
+      reducers,
+      {
+        metaReducers,
+        runtimeChecks:
+        {
+          strictStateImmutability: true,
+          strictActionImmutability: true
+        }
+      }
+    ),
+    !environment.production ? StoreDevtoolsModule.instrument() : [],
+    EffectsModule.forRoot([LeagueEffects])
   ],
   providers: [
     AuthGuard,
